requirements.txt
PyYAML==6.0.1
pymongo==4.5.0
requests==2.31.0
